fix(Claim): remove `claim.parts_from_resource`
`claim.parts_from_resource` was provoking errors in equality calcuations of Claims, depending if they had requested or not that field. This made the Claim effectively not immutable. The small performance gains are not deemed good enough for this. Removing the property and replacing it with specific calls when required.

BREAKING CHANGE: Claim#parts_from_resource is removed.